# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Golf Pool App API
  version: 1.0.0
  description: |
    API for managing golf pools, teams, tournaments, and leaderboards.

    All endpoints (except `/health`) require a Firebase Auth ID token passed as a Bearer token.
    Admin-only endpoints additionally require the authenticated user to have `admin: true` in Firestore.

    Make sure to use the correct server URL depending on your environment:
    - Local development: `http://127.0.0.1:5001/golf-pool-app-492300/us-central1/api`
    - Production: `https://us-central1-golf-pool-app-492300.cloudfunctions.net/api`
servers:
  - url: http://127.0.0.1:5001/golf-pool-app-492300/us-central1/api
    description: Local emulator
  - url: https://us-central1-golf-pool-app-492300.cloudfunctions.net/api
    description: Production (Firebase Cloud Function)s
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Firebase Auth ID token
  schemas:
    TierConfig:
      type: object
      required: [tierNumber, oddsMin, oddsMax, picksRequired]
      properties:
        tierNumber:
          type: integer
          example: 1
        oddsMin:
          type: string
          example: "+100"
        oddsMax:
          type: string
          example: "+500"
        picksRequired:
          type: integer
          example: 2
    ScoringRule:
      type: object
      required: [countBest, outOf]
      properties:
        countBest:
          type: integer
          description: Count the best N scores
          example: 4
        outOf:
          type: integer
          description: Out of the total picks per team
          example: 6
...
    PlayerScore:
      type: object
      properties:
        playerId:
          type: string
        playerName:
          type: string
        score:
          type: integer
          nullable: true
          description: Score relative to par (null if not started)
        missedCut:
          type: boolean
        counting:
          type: boolean
          description: Whether this score counts toward the team total
    LeaderboardEntry:
      type: object
      properties:
        teamId:
          type: string
        teamName:
          type: string
        userId:
          type: string
        totalScore:
          type: integer
          description: Combined score (lower is better)
        playerScores:
          type: array
          items:
            $ref: "#/components/schemas/PlayerScore"
...
    PoolSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        tournamentId:
          type: string
        tournamentName:
          type: string
          nullable: true
        tournamentStatus:
          type: string
          enum: [upcoming, active, completed]
          nullable: true
        createdBy:
          type: string
        myTeamId:
          type: string
          nullable: true
          description: The caller's team ID in this pool, or null if they have no team yet
    PoolDetail:
      allOf:
        - $ref: "#/components/schemas/PoolSummary"
        - type: object
          properties:
            tiers:
              type: array
              items:
                $ref: "#/components/schemas/TierConfig"
            scoringRule:
              $ref: "#/components/schemas/ScoringRule"
    UpcomingTeam:
      type: object
      properties:
        teamId:
          type: string
        teamName:
          type: string
        userId:
          type: string
        players:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              odds:
                type: string
    Error:
      type: object
      properties:
        error:
          type: string
      example:
        error: "Resource not found"
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid auth token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadGateway:
      description: Upstream ESPN API error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    SuccessResponse:
      description: Operation succeeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Success"
security:
  - BearerAuth: []
...
paths:
...
  /pools:
    post:
      summary: Create a pool
      description: Creates a new pool linked to a tournament. The authenticated user becomes the pool creator.
      tags: [Pools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, password, tournamentId, tiers, scoringRule]
              properties:
                name:
                  type: string
                  example: "Friends League 2026"
                password:
                  type: string
                  example: "birdie42"
                tournamentId:
                  type: string
                  example: "abc123"
                tiers:
                  type: array
                  items:
                    $ref: "#/components/schemas/TierConfig"
                scoringRule:
                  $ref: "#/components/schemas/ScoringRule"
            example:
              name: "Friends League 2026"
              password: "birdie42"
              tournamentId: "abc123"
              tiers:
                - tierNumber: 1
                  oddsMin: "+100"
                  oddsMax: "+500"
                  picksRequired: 2
                - tierNumber: 2
                  oddsMin: "+501"
                  oddsMax: "+2000"
                  picksRequired: 2
                - tierNumber: 3
                  oddsMin: "+2001"
                  oddsMax: "+9999"
                  picksRequired: 2
              scoringRule:
                countBest: 4
                outOf: 6
      responses:
        "201":
          description: Pool created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
              example:
                id: "pool_abc"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Tournament not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /pools/mine:
    get:
      summary: List pools the caller belongs to
      description: |
        Returns pools where the caller either has a team or is the creator.
        Each entry includes `myTeamId` (null if the caller hasn't created a team yet in that pool).
      tags: [Pools]
      responses:
        "200":
          description: Pool list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PoolSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /pools/{poolId}:
    get:
      summary: Get pool detail
      description: |
        Returns pool configuration (tiers, scoring rule, tournament summary, myTeamId).
        Does NOT return the pool password. Required by the team picker and pool detail screens.
      tags: [Pools]
      parameters:
        - in: path
          name: poolId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pool detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PoolDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /pools/{poolId}/join:
    post:
      summary: Join a pool
      description: |
        Validates the pool password. On success, the user should then call
        `POST /teams/pools/{poolId}/teams` to create their team.
      tags: [Pools]
      parameters:
        - in: path
          name: poolId
          required: true
          schema:
            type: string
          example: "pool_abc"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
                  example: "birdie42"
      responses:
        "200":
          description: Password accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  poolId:
                    type: string
                  message:
                    type: string
              example:
                success: true
                poolId: "pool_abc"
                message: "Password accepted. Create a team to join."
        "401":
          description: Incorrect pool password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /pools/{poolId}/leave:
    post:
      summary: Leave a pool
      description: Deletes the authenticated user's team from the pool.
      tags: [Pools]
      parameters:
        - in: path
          name: poolId
          required: true
          schema:
            type: string
          example: "pool_abc"
      responses:
        "200":
          $ref: "#/components/responses/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: User does not have a team in this pool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /pools/{poolId}/leaderboard:
    get:
      summary: Get the pool leaderboard
      description: |
        Returns all teams in the pool with scores.

        - If the tournament status is `upcoming`, returns teams with player info but no scores.
        - If the tournament is `active` or `completed`, returns live/final scores from ESPN, sorted ascending (lower = better).
      tags: [Pools]
      parameters:
        - in: path
          name: poolId
          required: true
          schema:
            type: string
          example: "pool_abc"
      responses:
        "200":
          description: Leaderboard data
          content:
            application/json:
              schema:
                oneOf:
                  - title: Upcoming Tournament
                    type: object
                    properties:
                      status:
                        type: string
                        enum: [upcoming]
                      teams:
                        type: array
                        items:
                          $ref: "#/components/schemas/UpcomingTeam"
                  - title: Active or Completed Tournament
                    type: object
                    properties:
                      status:
                        type: string
                        enum: [active, completed]
                      leaderboard:
                        type: array
                        description: Sorted ascending by totalScore (lowest wins)
                        items:
                          $ref: "#/components/schemas/LeaderboardEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "502":
          $ref: "#/components/responses/BadGateway"

  # ─────────────────────────────────────────────
  # Teams
  # ─────────────────────────────────────────────
...