# functions/src/routes/pools.ts
import { Router } from "express";
import { db } from "../config/firebase";
import { AuthRequest, requireAuth } from "../middleware/auth";
import { FieldValue } from "firebase-admin/firestore";
import {
  fetchScoreboard,
  fetchScoreboardForEvent,
} from "../services/espn";
import {
  parseScore,
  applyMissedCutPenalty,
  calculateTeamScore,
} from "../services/leaderboard";
import { LeaderboardEntry } from "../types";

const router = Router();

// POST /pools — Create pool (auth required)
router.post("/", requireAuth, async (req: AuthRequest, res) => {
  const { name, password, tournamentId, tiers, scoringRule } = req.body;

  if (!name || !password || !tournamentId || !tiers || !scoringRule) {
    res.status(400).json({
      error: "name, password, tournamentId, tiers, and scoringRule are required",
    });
    return;
  }

  // Verify tournament exists
  const tournDoc = await db.collection("tournaments").doc(tournamentId).get();
  if (!tournDoc.exists) {
    res.status(404).json({ error: "Tournament not found" });
    return;
  }

  const doc = await db.collection("pools").add({
    name,
    password,
    tournamentId,
    createdBy: req.uid,
    tiers,
    scoringRule,
    createdAt: FieldValue.serverTimestamp(),
    updatedAt: FieldValue.serverTimestamp(),
  });

  res.status(201).json({ id: doc.id });
});

// POST /pools/:poolId/join — Join pool
router.post("/:poolId/join", requireAuth, async (req: AuthRequest, res) => {
  const { poolId } = req.params;
  const { password } = req.body;

  const poolDoc = await db.collection("pools").doc(poolId).get();
  if (!poolDoc.exists) {
    res.status(404).json({ error: "Pool not found" });
    return;
  }

  const pool = poolDoc.data()!;
  if (pool.password !== password) {
    res.status(401).json({ error: "Incorrect password" });
    return;
  }

  // Check if user already has a team in this pool
  const existing = await db
    .collection("teams")
    .where("poolId", "==", poolId)
    .where("userId", "==", req.uid)
    .limit(1)
    .get();

  if (!existing.empty) {
    res.status(409).json({ error: "You already have a team in this pool" });
    return;
  }

  res.json({ success: true, poolId, message: "Password accepted. Create a team to join." });
});

// POST /pools/:poolId/leave — Leave pool
router.post("/:poolId/leave", requireAuth, async (req: AuthRequest, res) => {
  const { poolId } = req.params;

  // Find and delete user's team in this pool
  const teamsSnap = await db
    .collection("teams")
    .where("poolId", "==", poolId)
    .where("userId", "==", req.uid)
    .get();

  if (teamsSnap.empty) {
    res.status(404).json({ error: "You don't have a team in this pool" });
    return;
  }

  const batch = db.batch();
  teamsSnap.docs.forEach((doc) => batch.delete(doc.ref));
  await batch.commit();

  res.json({ success: true });
});

// GET /pools/mine — List pools the caller is a member of (has a team) or created.
// Registered before /:poolId routes so Express doesn't treat "mine" as a poolId.
router.get("/mine", requireAuth, async (req: AuthRequest, res) => {
  // Pools where I have a team → capture myTeamId per pool
  const teamsSnap = await db
    .collection("teams")
    .where("userId", "==", req.uid)
    .get();
  const teamIdByPool = new Map<string, string>();
  teamsSnap.docs.forEach((d) => {
    teamIdByPool.set(d.data().poolId as string, d.id);
  });

  // Pools I created (even if I haven't made a team yet)
  const createdSnap = await db
    .collection("pools")
    .where("createdBy", "==", req.uid)
    .get();

  const poolIds = new Set<string>([
    ...teamIdByPool.keys(),
    ...createdSnap.docs.map((d) => d.id),
  ]);

  if (poolIds.size === 0) {
    res.json([]);
    return;
  }

  // Fetch pool docs + tournament docs (dedupe tournament lookups)
  const poolDocs = await Promise.all(
    [...poolIds].map((id) => db.collection("pools").doc(id).get()),
  );
  const tournamentIds = new Set<string>();
  poolDocs.forEach((p) => {
    if (p.exists) tournamentIds.add(p.data()!.tournamentId as string);
  });
  const tournamentDocs = await Promise.all(
    [...tournamentIds].map((id) => db.collection("tournaments").doc(id).get()),
  );
  const tournamentById = new Map<
    string,
    { name: string; status: "upcoming" | "active" | "completed" }
  >();
  tournamentDocs.forEach((t) => {
    if (t.exists) {
      const data = t.data()!;
      tournamentById.set(t.id, { name: data.name, status: data.status });
    }
  });

  const result = poolDocs
    .filter((p) => p.exists)
    .map((p) => {
      const data = p.data()!;
      const tourn = tournamentById.get(data.tournamentId);
      return {
        id: p.id,
        name: data.name,
        tournamentId: data.tournamentId,
        tournamentName: tourn?.name ?? null,
        tournamentStatus: tourn?.status ?? null,
        createdBy: data.createdBy,
        myTeamId: teamIdByPool.get(p.id) ?? null,
      };
    });

  res.json(result);
});

// GET /pools/:poolId — Pool detail (tiers, scoring rule, tournament summary).
// Does NOT return the password. Used by the team picker and pool detail screens.
router.get("/:poolId", requireAuth, async (req: AuthRequest, res) => {
  const { poolId } = req.params;

  const poolDoc = await db.collection("pools").doc(poolId).get();
  if (!poolDoc.exists) {
    res.status(404).json({ error: "Pool not found" });
    return;
  }
  const pool = poolDoc.data()!;

  const tournDoc = await db
    .collection("tournaments")
    .doc(pool.tournamentId)
    .get();
  const tournament = tournDoc.exists ? tournDoc.data()! : null;

  // myTeamId for the caller (if any)
  const teamSnap = await db
    .collection("teams")
    .where("poolId", "==", poolId)
    .where("userId", "==", req.uid)
    .limit(1)
    .get();
  const myTeamId = teamSnap.empty ? null : teamSnap.docs[0].id;

  res.json({
    id: poolDoc.id,
    name: pool.name,
    tournamentId: pool.tournamentId,
    tournamentName: tournament?.name ?? null,
    tournamentStatus: tournament?.status ?? null,
    createdBy: pool.createdBy,
    tiers: pool.tiers,
    scoringRule: pool.scoringRule,
    myTeamId,
  });
});

// GET /pools/:poolId/leaderboard — Get leaderboard
router.get("/:poolId/leaderboard", requireAuth, async (req: AuthRequest, res) => {
  const { poolId } = req.params;

  const poolDoc = await db.collection("pools").doc(poolId).get();
  if (!poolDoc.exists) {
    res.status(404).json({ error: "Pool not found" });
    return;
  }
  const pool = poolDoc.data()!;

  // Get tournament
  const tournDoc = await db.collection("tournaments").doc(pool.tournamentId).get();
  if (!tournDoc.exists) {
    res.status(404).json({ error: "Tournament not found" });
    return;
  }
  const tournament = tournDoc.data()!;

  // Get all teams in pool
  const teamsSnap = await db
    .collection("teams")
    .where("poolId", "==", poolId)
    .get();

  if (teamsSnap.empty) {
    res.json({ leaderboard: [] });
    return;
  }

  // If tournament hasn't started, return teams without scores
  if (tournament.status === "upcoming") {
    const teams = await Promise.all(
      teamsSnap.docs.map(async (teamDoc) => {
        const team = teamDoc.data();
        const playerDocs = await Promise.all(
          team.picks.map((id: string) => db.collection("players").doc(id).get())
        );
        return {
          teamId: teamDoc.id,
          teamName: team.name,
          userId: team.userId,
          players: playerDocs
            .filter((d) => d.exists)
            .map((d) => ({ id: d.id, name: d.data()!.name, odds: d.data()!.odds })),
        };
      })
    );
    res.json({ status: "upcoming", teams });
    return;
  }

  // Tournament is active or completed — fetch ESPN scores
  let competitors;
  try {
    competitors = tournament.espnEventId
      ? await fetchScoreboardForEvent(tournament.espnEventId)
      : await fetchScoreboard();
  } catch {
    res.status(502).json({ error: "Failed to fetch ESPN scores" });
    return;
  }

  // Build score map: espnId -> { score, missedCut }
  const scoreMap = new Map<string, { score: number; missedCut: boolean }>();
  for (const c of competitors) {
    const missedCut =
      c.status?.type?.name === "STATUS_CUT" ||
      c.status?.type?.description?.toLowerCase().includes("cut");
    scoreMap.set(c.id, {
      score: parseScore(c.score),
      missedCut: !!missedCut,
    });
  }

  // Apply missed-cut penalty
  const adjustedScores = applyMissedCutPenalty(scoreMap);

  // Calculate each team's score
  const leaderboard: LeaderboardEntry[] = await Promise.all(
    teamsSnap.docs.map(async (teamDoc) => {
      const team = teamDoc.data();
      const playerDocs = await Promise.all(
        team.picks.map((id: string) => db.collection("players").doc(id).get())
      );

      const playerScores = playerDocs
        .filter((d) => d.exists)
        .map((d) => {
          const data = d.data()!;
          const espnData = data.espnId ? adjustedScores.get(data.espnId) : null;
          return {
            playerId: d.id,
            playerName: data.name,
            espnId: data.espnId,
            score: espnData?.score ?? null,
            missedCut: espnData?.missedCut ?? false,
          };
        });

      const result = calculateTeamScore(playerScores, pool.scoringRule);
      return {
        teamId: teamDoc.id,
        teamName: team.name,
        userId: team.userId,
        totalScore: result.totalScore,
        playerScores: result.map((r) => ({
          playerId: r.playerId,
          playerName: r.playerName,
          score: r.score,
          missedCut: r.missedCut,
          counting: r.counting,
        })),
      };
    })
  );

  // Sort by total score ascending (lowest is best in golf)
  leaderboard.sort((a, b) => a.totalScore - b.totalScore);

  res.json({ status: tournament.status, leaderboard });
});

export default router;